wallet: remove withdraw saga
