+++ app/containers/WalletPage/saga.js
@@ -1,17 +1,10 @@
 import { all, call, put, select, takeLatest } from 'redux-saga/effects';
-import { LOAD_TRANSACTIONS, LOAD_WITHDRAW } from './constants';
+import { LOAD_TRANSACTIONS } from './constants';
 import { makeSelectCurrentUser } from '../App/selectors';
 import api from '../../utils/barter-dex-api';
-import {
-  loadTransactionsSuccess,
-  loadTransactionsError,
-  loadWithdrawSuccess,
-  loadWithdrawError
-} from './actions';
+import { loadTransactionsSuccess, loadTransactionsError } from './actions';
 
 const debug = require('debug')('dicoapp:containers:WalletPage:saga');
-
-const numcoin = 100000000;
 
 export function* loadCoinTransactionsProcess(coin, address, userpass) {
   try {
@@ -69,53 +62,9 @@ export function* loadTransactionsProcess() {
   }
 }
 
-export function* loadWithdrawProcess({ payload }) {
-  try {
-    // load user data
-    const user = yield select(makeSelectCurrentUser());
-    if (!user) {
-      throw new Error('not found user');
-    }
-    const userpass = user.get('userpass');
-
-    const { amount, address, coin } = payload;
-
-    let outputs = `[{${address}: ${Number(amount)}}]`;
-
-    // eslint-disable-next-line no-eval
-    outputs = JSON.stringify(eval(`(${outputs})`));
-
-    const sendparams = {
-      userpass,
-      coin,
-      outputs: JSON.parse(outputs)
-    };
-
-    const resultWithdraw = yield api.withdraw(sendparams);
-
-    const { hex, txfee } = resultWithdraw;
-
-    const sendrawtx = {
-      userpass,
-      coin,
-      signedtx: hex
-    };
-    const resultSendrawtx = yield api.sendRawTransaction(sendrawtx);
-    debug(`resultSendrawtx = ${resultSendrawtx}`);
-
-    // eslint-disable-next-line no-param-reassign
-    payload.amount += txfee / numcoin;
-
-    return yield put(loadWithdrawSuccess(payload));
-  } catch (err) {
-    return yield put(loadWithdrawError(err.message));
-  }
-}
-
 /**
  * Root saga manages watcher lifecycle
  */
 export default function* walletData() {
   yield takeLatest(LOAD_TRANSACTIONS, loadTransactionsProcess);
-  yield takeLatest(LOAD_WITHDRAW, loadWithdrawProcess);
 }
